remove stale connections

+++ examples/uwb_messages.h
@@ -31,9 +31,9 @@ typedef struct
     uint32 seat_map;
     uint8 rx_ts[4];
     uint8 tx_ts[4];
-    uint32 x_pos;
-    uint32 y_pos;
-    uint32 z_pos;
+    double x_pos;
+    double y_pos;
+    double z_pos;
     uint8 data_type;
     uint8 data1[4];
     uint8 data2[4];
@@ -56,7 +56,7 @@ typedef struct
     uint8 seat_num;
     uint8 timeout_ctr;
     double distance;
-    uint32 x_pos;
-    uint32 y_pos;
-    uint32 z_pos;
+    double x_pos;
+    double y_pos;
+    double z_pos;
 } uwb_device_t;
